Created files for different components
